Constant time class to class ancestor lookup
Previously when checking ancestors, we would walk all the way up the
ancestry chain checking each parent for a matching class or module.

I believe this was especially unfriendly to CPU cache since for each
step we need to check two cache lines (the class and class ext). We also
probably weren't allowing the CPU any speculation through this linked
list.

This check is used quite often in:
* case statements
* Calling protected methods
* Class#is_a?
* Module#===
* Module#<=>

I believe it's most common to check a class against a parent class, to
this commit aims to improve that (unfortunately does not help checking
for a non-class parent).

This is done by storing on each class the number and an array of all
parent classes, in order. Using this we can check whether a class is a
subclass of another in constant time since we know the location to
expect it in the hierarchy.

This does not implement the optimization for singleton classes, since
that seemed to be less common and had more complexity.