Slightly better numeric casting
Casting to usize once at the top removes one cast. We also did not check
that argc is non negative before this change. I guess the C code is
using `int` for the field even though it's never negative.

Feel free to revert / make changes if you disagree that this is an
improvement.