Take VM lock from Rust side. Move branch_stub_hit() to Rust
A lot of APIs we expose in yjit.h need the VM lock for ractor support
so being able to take the lock from the Rust side will be helpful.

Defining branch_stub_hit() in Rust lets us use LLVM to generate
a function with the SysV calling convention even on Windows. It's only
called from generated code so this keeps the generator portable.