Use a HashSet<BlockRef> for a few invariants like in C YJIT
It's important that we are able to quickly remove a block from the
Invariants struct and keeping the BlockRefs in a Vec meant we need
a linear search for the removal.

Implement by-identity Hash and Eq for BlockRef and use HashSet for
`single_ractor`, `basic_operators`, and `global_constant_state`.
This is like in C YJIT, where we use `st_table` as a hash set for
the blocks.