Sever dependency on Cargo in release mode
We want to build offline in release mode and currently `cargo --offline`
does not do what we want. In release mode, run `rustc` directly and
don't check for Cargo in `configure`. Reword the build prompt to include
the current build mode.

No CI changes yet, but this can be used to check that we build without
any external crates. For release mode users it's potentially a boon to
not need to install Cargo.