Support buildir!=srcdir build in bindgen. Run on CI
This swaps test-bundler-parallel with the bindgen check. The Bundler
tests are relatively slow which I suspect is why they were removed
upstream.

I had to name an unnamed enum to get stable bindgen output on CI.
With the unnamed enum rust-bindgen picks an arbitrary name that is
different maybe depending on the libclang version.