Memoize match cache points on backtracking
The previous implementation records a match cache point on arrived this.
However, such a implementation introduced the null-loop problem; that
is, null-loops that should be matched are recorded and fail to match.
When we memoize (record) cache points on backtracking, we can avoid this
annoying problem.

Additionally, we expect this implementation makes to support look-around
operators available.